Refactor and clean up of shared
