+++ product_code/lagom/authentication_service/impl/src/main/scala/de/upb/cs/uc4/authentication/impl/AuthenticationSerializerRegistry.scala
@@ -1,7 +1,7 @@
 package de.upb.cs.uc4.authentication.impl
 
 import com.lightbend.lagom.scaladsl.playjson.JsonSerializer
-import de.upb.cs.uc4.shared.SharedSerializerRegistry
+import de.upb.cs.uc4.shared.server.SharedSerializerRegistry
 
 object AuthenticationSerializerRegistry extends SharedSerializerRegistry {
   /** All Service specific JsonSerializers */
