+++ product_code/lagom/course_service/impl/src/main/scala/de/upb/cs/uc4/course/impl/CourseSerializerRegistry.scala
@@ -3,7 +3,7 @@ package de.upb.cs.uc4.course.impl
 import com.lightbend.lagom.scaladsl.playjson.JsonSerializer
 import de.upb.cs.uc4.course.impl.actor.CourseState
 import de.upb.cs.uc4.course.impl.events.{OnCourseCreate, OnCourseDelete, OnCourseUpdate}
-import de.upb.cs.uc4.shared.SharedSerializerRegistry
+import de.upb.cs.uc4.shared.server.SharedSerializerRegistry
 
 import scala.collection.immutable.Seq
