+++ product_code/lagom/course_service/impl/src/main/scala/de/upb/cs/uc4/course/impl/commands/CreateCourse.scala
@@ -2,6 +2,6 @@ package de.upb.cs.uc4.course.impl.commands
 
 import akka.actor.typed.ActorRef
 import de.upb.cs.uc4.course.model.Course
-import de.upb.cs.uc4.shared.messages.Confirmation
+import de.upb.cs.uc4.shared.server.messages.Confirmation
 
 case class CreateCourse(course: Course, replyTo: ActorRef[Confirmation]) extends CourseCommand
