+++ product_code/lagom/course_service/impl/src/main/scala/de/upb/cs/uc4/course/impl/commands/DeleteCourse.scala
@@ -1,6 +1,6 @@
 package de.upb.cs.uc4.course.impl.commands
 
 import akka.actor.typed.ActorRef
-import de.upb.cs.uc4.shared.messages.Confirmation
+import de.upb.cs.uc4.shared.server.messages.Confirmation
 
 case class DeleteCourse(id: String, replyTo: ActorRef[Confirmation]) extends CourseCommand
