+++ product_code/lagom/hyperledger_service/impl/src/main/scala/de/upb/cs/uc4/hyperledger/impl/HyperLedgerSerializerRegistry.scala
@@ -1,7 +1,7 @@
 package de.upb.cs.uc4.hyperledger.impl
 
 import com.lightbend.lagom.scaladsl.playjson.JsonSerializer
-import de.upb.cs.uc4.shared.SharedSerializerRegistry
+import de.upb.cs.uc4.shared.server.SharedSerializerRegistry
 
 import scala.collection.immutable.Seq
