+++ product_code/lagom/hyperledger_service/impl/src/main/scala/de/upb/cs/uc4/hyperledger/impl/actor/HyperLedgerBehaviour.scala
@@ -6,7 +6,7 @@ import akka.cluster.sharding.typed.scaladsl.EntityTypeKey
 import de.upb.cs.uc4.hyperledger.impl.HyperLedgerApplication
 import de.upb.cs.uc4.hyperledger.impl.commands.{HyperLedgerCommand, Read, Shutdown, Write}
 import de.upb.cs.uc4.hyperledger.traits.{ChaincodeTrait, ConnectionManagerTrait}
-import de.upb.cs.uc4.shared.messages.{Accepted, Rejected}
+import de.upb.cs.uc4.shared.server.messages.{Accepted, Rejected}
 
 object HyperLedgerBehaviour {
