+++ product_code/lagom/hyperledger_service/impl/src/main/scala/de/upb/cs/uc4/hyperledger/impl/commands/Write.scala
@@ -1,6 +1,6 @@
 package de.upb.cs.uc4.hyperledger.impl.commands
 
 import akka.actor.typed.ActorRef
-import de.upb.cs.uc4.shared.messages.Confirmation
+import de.upb.cs.uc4.shared.server.messages.Confirmation
 
 case class Write(json: String, replyTo: ActorRef[Confirmation]) extends HyperLedgerCommand
