+++ product_code/lagom/shared/client/src/main/scala/de/upb/cs/uc4/shared/client/CustomException.scala
@@ -1,0 +1,18 @@
+package de.upb.cs.uc4.shared.client
+
+import com.lightbend.lagom.scaladsl.api.transport.TransportErrorCode
+
+
+class CustomException(errorCode: TransportErrorCode, possibleErrorResponse: DetailedError, cause: Throwable) extends Exception(possibleErrorResponse.title, null, true, true){
+
+  def this(errorCode: TransportErrorCode, possibleErrorResponse: DetailedError) =
+    this(errorCode, possibleErrorResponse, null)
+
+
+  def getErrorCode: TransportErrorCode = {
+    errorCode
+  }
+  def getPossibleErrorResponse: DetailedError = {
+    possibleErrorResponse
+  }
+}
