+++ product_code/lagom/shared/client/src/main/scala/de/upb/cs/uc4/shared/client/SimpleError.scala
@@ -1,0 +1,9 @@
+package de.upb.cs.uc4.shared.client
+
+import play.api.libs.json.{Format, Json}
+
+case class SimpleError(name: String, reason: String)
+
+object SimpleError {
+  implicit val format: Format[SimpleError] = Json.format
+}
