+++ product_code/lagom/shared/server/src/main/scala/de/upb/cs/uc4/shared/server/SharedSerializerRegistry.scala
@@ -1,0 +1,18 @@
+package de.upb.cs.uc4.shared.server
+
+import com.lightbend.lagom.scaladsl.playjson.{JsonSerializer, JsonSerializerRegistry}
+import de.upb.cs.uc4.shared.server.messages.{Accepted, Confirmation, Rejected}
+
+import scala.collection.immutable.Seq
+
+/** Automatically creates Serializers for all shared messages, events etc */
+trait SharedSerializerRegistry extends JsonSerializerRegistry {
+  override def serializers: Seq[JsonSerializer[_]] = Seq(
+    JsonSerializer[Confirmation],
+    JsonSerializer[Accepted],
+    JsonSerializer[Rejected]
+  ) ++ customSerializers
+
+  /** All Service specific JsonSerializers */
+  def customSerializers: Seq[JsonSerializer[_]]
+}
