--- product_code/lagom/shared/src/main/scala/de/upb/cs/uc4/shared/messages/Accepted.scala
@@ -1,11 +1,0 @@
-package de.upb.cs.uc4.shared.messages
-
-import play.api.libs.json._
-
-/** Accepted version of the Confirmation */
-trait Accepted extends Confirmation
-
-case object Accepted extends Accepted {
-  implicit val format: Format[Accepted] =
-    Format(Reads(_ => JsSuccess(Accepted)), Writes(_ => Json.obj()))
-}
