--- product_code/lagom/shared/src/main/scala/de/upb/cs/uc4/shared/messages/Rejected.scala
@@ -1,10 +1,0 @@
-package de.upb.cs.uc4.shared.messages
-
-import play.api.libs.json.{Format, Json}
-
-/** Rejected version of the Confirmation */
-case class Rejected(reason: String) extends Confirmation
-
-object Rejected {
-  implicit val format: Format[Rejected] = Json.format
-}
