--- product_code/lagom/shared/src/main/scala/de/upb/cs/uc4/shared/messages/RejectedWithError.scala
@@ -1,11 +1,0 @@
-package de.upb.cs.uc4.shared.messages
-
-import de.upb.cs.uc4.shared.api.DetailedError
-import play.api.libs.json.{Format, Json}
-
-/** RejectedWithError version of the Confirmation */
-case class RejectedWithError(statusCode: Int, reason: DetailedError) extends Confirmation
-
-object RejectedWithError {
-  implicit val format: Format[RejectedWithError] = Json.format
-}
