+++ product_code/lagom/user_service/impl/src/main/scala/de/upb/cs/uc4/user/impl/UserSerializerRegistry.scala
@@ -1,7 +1,7 @@
 package de.upb.cs.uc4.user.impl
 
 import com.lightbend.lagom.scaladsl.playjson.JsonSerializer
-import de.upb.cs.uc4.shared.SharedSerializerRegistry
+import de.upb.cs.uc4.shared.server.SharedSerializerRegistry
 import de.upb.cs.uc4.user.impl.actor.UserState
 import de.upb.cs.uc4.user.impl.events.{OnUserCreate, OnUserDelete, OnUserUpdate}
