+++ product_code/lagom/user_service/impl/src/main/scala/de/upb/cs/uc4/user/impl/actor/UserState.scala
@@ -3,9 +3,9 @@ package de.upb.cs.uc4.user.impl.actor
 
 import akka.cluster.sharding.typed.scaladsl.EntityTypeKey
 import akka.persistence.typed.scaladsl.{Effect, ReplyEffect}
-import de.upb.cs.uc4.shared.api
-import de.upb.cs.uc4.shared.api.{DetailedError, SimpleError}
-import de.upb.cs.uc4.shared.messages._
+import de.upb.cs.uc4.shared.client
+import de.upb.cs.uc4.shared.client.{DetailedError, SimpleError}
+import de.upb.cs.uc4.shared.server.messages._
 import de.upb.cs.uc4.user.impl.UserApplication
 import de.upb.cs.uc4.user.impl.commands._
 import de.upb.cs.uc4.user.impl.events.{OnUserCreate, OnUserDelete, OnUserUpdate, UserEvent}
@@ -49,7 +49,7 @@ case class UserState(optUser: Option[User]) {
           if(validationErrors.isEmpty){
             Effect.persist(OnUserUpdate(user)).thenReply(replyTo) { _ => Accepted }
           } else {
-            Effect.reply(replyTo)(RejectedWithError(422, api.DetailedError("validation error", "Your request parameters did not validate", validationErrors)))
+            Effect.reply(replyTo)(RejectedWithError(422, client.DetailedError("validation error", "Your request parameters did not validate", validationErrors)))
           }
         } else {
           Effect.reply(replyTo)(RejectedWithError(404, DetailedError("key value error", "Username not found", List(SimpleError("username", "Username does not exist")))))
