+++ product_code/lagom/user_service/impl/src/main/scala/de/upb/cs/uc4/user/impl/commands/DeleteUser.scala
@@ -1,6 +1,6 @@
 package de.upb.cs.uc4.user.impl.commands
 
 import akka.actor.typed.ActorRef
-import de.upb.cs.uc4.shared.messages.Confirmation
+import de.upb.cs.uc4.shared.server.messages.Confirmation
 
 case class DeleteUser(replyTo: ActorRef[Confirmation]) extends UserCommand
