+++ product_code/lagom/user_service/impl/src/main/scala/de/upb/cs/uc4/user/impl/commands/UpdateUser.scala
@@ -1,7 +1,7 @@
 package de.upb.cs.uc4.user.impl.commands
 
 import akka.actor.typed.ActorRef
-import de.upb.cs.uc4.shared.messages.Confirmation
+import de.upb.cs.uc4.shared.server.messages.Confirmation
 import de.upb.cs.uc4.user.impl.actor.User
 
 case class UpdateUser(user: User, replyTo: ActorRef[Confirmation]) extends UserCommand
